break out types and rename them also
